feat(web-monitor-server-golang): 完成定时任务的部分功能（未解决循环依赖的问题）

+++ web-monitor-server/web-monitor-server-golang/model/validation/scheduler.go
@@ -6,10 +6,10 @@ type SchedulerRule struct {
 }
 
 type SchedulerRuleItem struct {
-	Agg      string `json:"agg"`
-	Ind      string `json:"ind"`
-	Interval int    `json:"interval"`
-	Op       string `json:"op"`
-	TimeSpan int    `json:"timeSpan"`
-	Val      int    `json:"val"`
+	Agg      string  `json:"agg"`
+	Ind      string  `json:"ind"`
+	Interval int     `json:"interval"`
+	Op       string  `json:"op"`
+	TimeSpan int     `json:"timeSpan"`
+	Val      float64 `json:"val"`
 }
